astraweave_ecs::events

Struct Events

pub struct Events { /* private fields */ }

Central event registry for all event types

Implementations

impl Events

pub fn new() -> Self

Creates a new event registry with the default retention of 2 frames.

Examples

use astraweave_ecs::Events;

let events = Events::new();

pub fn with_keep_frames(self, frames: u64) -> Self

pub fn keep_frames(&self) -> u64

Get the number of frames events are kept before cleanup.

pub fn send<E: Event>(&mut self, event: E)

Send an event of type E.

The event is queued and can be read by any EventReader<E> until it is cleaned up after the configured retention period.

Examples

use astraweave_ecs::{Event, Events};

struct DamageEvent { amount: i32 }
impl Event for DamageEvent {}

let mut events = Events::new();
events.send(DamageEvent { amount: 10 });

pub fn get_reader<E: Event>(&self) -> EventReader<E>

Get event reader for type E

pub fn read<E: Event>(&self) -> impl Iterator<Item = &E>

Read events of type E

pub fn drain<E: Event>(&mut self) -> impl Iterator<Item = E> + '_

Drain all events of type E (consumes them)

pub fn clear<E: Event>(&mut self)

Clear all events of type E

pub fn len<E: Event>(&self) -> usize

Get event count for type E

pub fn is_empty<E: Event>(&self) -> bool

Check if events queue is empty for type E

pub fn update(&mut self)

Advance frame and cleanup old events

pub fn clear_all(&mut self)

Clear all events

pub fn current_frame(&self) -> u64